requirements.txt
torch>=1.9.0
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
PyYAML>=6.0
matplotlib>=3.4.0
seaborn>=0.11.0
jupyter>=1.0.0
